changes to simplesol
